LayoutTests/http/tests/fetch/response-clone-blob-expected.txt

Tests that calling blob() on a cloned Fetch Response does not disturb the original response.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS originalText == cloneText is true
PASS successfullyParsed is true

TEST COMPLETE
